Remove engine image dependency of v2 volumes
engineImage in the spec fields of volume, engine and replica are replaced with image,
and image can be either engine image or instance manager image.
Several existing logics and validation checks specifically identify the engine image.
The objective is to broaden these logics and checks to encompass both engine and
instance manager images.

Longhorn 7157

Signed-off-by: Derek Su <derek.su@suse.com>