disk: record the name of the instance manager managing it
A block-type disk can only be managed by a specific spdk_tgt within an
instance manager. The commit is intended to record the name of the
instance manager associated with that disk.

On the other hand, for a filesystem-type disk, any instance manager can
access it. Therefore, the recorded instance manager for a filesystem-type
disk is always the default instance manager.

Longhorn 6001

Signed-off-by: Derek Su <derek.su@suse.com>